+++ src/main/java/kitchenpos/order/domain/Order.java
@@ -20,8 +20,8 @@ public class Order {
     @Enumerated(EnumType.STRING)
     private OrderStatus orderStatus;
     private LocalDateTime orderedTime;
-    @OneToMany(mappedBy = "order", cascade = CascadeType.ALL, orphanRemoval = true)
-    private List<OrderLineItem> orderLineItems = new ArrayList<>();
+    @Embedded
+    OrderLineItems orderLineItems = new OrderLineItems();
 
     protected Order() {
     }
@@ -34,11 +34,10 @@ public class Order {
     }
 
     public Order(Long orderTableId, List<OrderLineItem> orderLineItems) {
-        validateOrderLineItems(orderLineItems);
         this.orderTableId = requireNonNull(orderTableId, "orderTableId");
         this.orderedTime = LocalDateTime.now();
         this.orderStatus = OrderStatus.COOKING;
-        this.orderLineItems = orderLineItems;
+        addOrderLineItems(orderLineItems);
     }
 
     public Order(Long orderTableId) {
@@ -46,23 +45,7 @@ public class Order {
     }
 
     public void addOrderLineItems(List<OrderLineItem> orderLineItems) {
-        validateOrderLineItems(orderLineItems);
-        for (OrderLineItem orderLineItem : orderLineItems) {
-            add(orderLineItem);
-        }
-    }
-
-    private void add(OrderLineItem orderLineItem) {
-        if (!this.orderLineItems.contains(orderLineItem)) {
-            this.orderLineItems.add(orderLineItem);
-        }
-        orderLineItem.bindTo(this);
-    }
-
-    private void validateOrderLineItems(List<OrderLineItem> orderLineItems) {
-        if (orderLineItems.isEmpty()) {
-            throw new IllegalArgumentException("주문내역 목록이 비어있습니다.");
-        }
+        this.orderLineItems.addAll(this, orderLineItems);
     }
 
     public void changeOrderStatus(OrderStatus orderStatus) {
@@ -93,7 +76,7 @@ public class Order {
     }
 
     public List<OrderLineItem> getOrderLineItems() {
-        return orderLineItems;
+        return orderLineItems.get();
     }
 
 }
